rename config.py to utils.py
