+++ pyLinguist/utils.py
